Share waveform fixtures
